Patch from Eric Warmenhoven <eric@warmenhoven.org>:
Fixes jde-parse-class.el to handle tableswitch opcode with the low and
high values "larger than emacs can handle".  Changes the
tableswitch handling so that instead of getting the low and high values
independently and subtracting, it will get both values at the same time
and do the subtraction in line with reconstructing a 4-byte value.