Patch submitted by "Jeff Peck" <jpjnk4-jdee@yahoo.com> and also covered by Sourceforge patch 2923583:
Make jde-lib-directory-names a little "looser". it used to match only
on the directory name component of the path, but this changes
jde-expand-classpath to match against the whole path, not just the
directory name.

a) it means things named: ".../jarx" or ".../lib2" are not seen as *.jar repositories
    (for backward compatibility, you could use: (list "/lib[^/]*$" "/jar[^/]*$") )
b) it means one can say: "/lib/ext$" *is* a jar repository
as in: '(jde-lib-directory-names '("/lib$" "/jar$" "/lib/ext$"))
which I need and could find no way to express, without including other foo/ext directories...
(and writing a jde-excluded-file-names to remove the extraneous files seems wrong;

 I expect the excluded list to operate on the 'tail' of the filename, not on the upper directories)