Apply fix from Troy Daniels (sourceforge tracker id 3154645):
jde-xref-get-db-directory normalizes the path by just passing the
value of the db directory. This gives the incorrect result when there
are multiple prj.el files and the one defining the variable is not the
deepest. The implementation should be

(concat (jde-normalize-path 'jde-xref-db-base-directory) "/xrefdb"))

instead of

(concat (jde-normalize-path jde-xref-db-base-directory) "/xrefdb"))

(Note that jde-xref-db-base-directory is quoted in the correct implementation.